reset:sft

+++ @build.sh
@@ -2,6 +2,11 @@
 
 # this file is used by "clone-all-org-repos" project
 
+if [[ "$1" != "--force" ]]; then
+ echo "are you sure you wish to rm -rf...? then use --force"
+ exit 1;
+fi
+
 cd $(dirname "$0") &&
 rm -rf node_modules &&
 npm install &&
